Refactor ad views
