+++ hybid.sdk/src/main/java/net/pubnative/lite/sdk/interstitial/HyBidInterstitialAd.java
@@ -1,0 +1,186 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2018 PubNative GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package net.pubnative.lite.sdk.interstitial;
+
+import android.app.Activity;
+import android.text.TextUtils;
+
+import net.pubnative.lite.sdk.api.InterstitialRequestManager;
+import net.pubnative.lite.sdk.api.RequestManager;
+import net.pubnative.lite.sdk.interstitial.presenter.InterstitialPresenter;
+import net.pubnative.lite.sdk.interstitial.presenter.InterstitialPresenterFactory;
+import net.pubnative.lite.sdk.models.Ad;
+import net.pubnative.lite.sdk.utils.Logger;
+
+public class HyBidInterstitialAd implements RequestManager.RequestListener, InterstitialPresenter.Listener {
+    private static final String TAG = HyBidInterstitialAd.class.getSimpleName();
+
+    public interface Listener {
+        void onInterstitialLoaded();
+
+        void onInterstitialLoadFailed(Throwable error);
+
+        void onInterstitialImpression();
+
+        void onInterstitialDismissed();
+
+        void onInterstitialClick();
+    }
+
+    private RequestManager mRequestManager;
+    private InterstitialPresenter mPresenter;
+    private Listener mListener;
+    private Activity mActivity;
+    private String mZoneId;
+    private boolean mReady = false;
+
+    public HyBidInterstitialAd(Activity activity, String zoneId, Listener listener) {
+        mRequestManager = new InterstitialRequestManager();
+        mActivity = activity;
+        mZoneId = zoneId;
+        mListener = listener;
+    }
+
+    public void load() {
+        if (TextUtils.isEmpty(mZoneId)) {
+            invokeOnLoadFailed(new Exception("Invalid zone id provided"));
+        } else {
+            cleanup();
+            mRequestManager.setZoneId(mZoneId);
+            mRequestManager.setRequestListener(this);
+            mRequestManager.requestAd();
+        }
+    }
+
+    public void show() {
+        if (mPresenter != null && mReady) {
+            mPresenter.show();
+        } else {
+            Logger.e(TAG, "Can't display ad. Interstitial not ready.");
+        }
+    }
+
+    public void hide() {
+        if (mPresenter != null) {
+            mPresenter.hide();
+        }
+    }
+
+    public boolean isReady() {
+        return mReady;
+    }
+
+    public void destroy() {
+        cleanup();
+        if (mRequestManager != null) {
+            mRequestManager.destroy();
+            mRequestManager = null;
+        }
+    }
+
+    private void cleanup() {
+        mReady = false;
+        if (mPresenter != null) {
+            mPresenter.destroy();
+            mPresenter = null;
+        }
+    }
+
+    private void renderAd(Ad ad) {
+        mPresenter = new InterstitialPresenterFactory(mActivity).createInterstitialPresenter(ad, this);
+        mPresenter.load();
+    }
+
+    protected void invokeOnLoadFinished() {
+        if (mListener != null) {
+            mListener.onInterstitialLoaded();
+        }
+    }
+
+    protected void invokeOnLoadFailed(Exception exception) {
+        Logger.e(TAG, exception.getMessage());
+        if (mListener != null) {
+            mListener.onInterstitialLoadFailed(exception);
+        }
+    }
+
+    protected void invokeOnClick() {
+        if (mListener != null) {
+            mListener.onInterstitialClick();
+        }
+    }
+
+    protected void invokeOnImpression() {
+        if (mListener != null) {
+            mListener.onInterstitialImpression();
+        }
+    }
+
+    protected void invokeOnDismissed() {
+        if (mListener != null) {
+            mListener.onInterstitialDismissed();
+        }
+    }
+
+    //------------------------------ RequestManager Callbacks --------------------------------------
+    @Override
+    public void onRequestSuccess(Ad ad) {
+        if (ad == null) {
+            invokeOnLoadFailed(new Exception("Server returned null ad"));
+        } else {
+            renderAd(ad);
+        }
+    }
+
+    @Override
+    public void onRequestFail(Throwable throwable) {
+        invokeOnLoadFailed(new Exception(throwable));
+    }
+
+    //------------------------- IntersititialPresenter Callbacks -----------------------------------
+    @Override
+    public void onInterstitialLoaded(InterstitialPresenter interstitialPresenter) {
+        mReady = true;
+        invokeOnLoadFinished();
+    }
+
+    @Override
+    public void onInterstitialError(InterstitialPresenter interstitialPresenter) {
+        invokeOnLoadFailed(new Exception("An error has occurred while rendering the interstitial"));
+    }
+
+    @Override
+    public void onInterstitialShown(InterstitialPresenter interstitialPresenter) {
+        invokeOnImpression();
+    }
+
+    @Override
+    public void onInterstitialClicked(InterstitialPresenter interstitialPresenter) {
+        invokeOnClick();
+    }
+
+    @Override
+    public void onInterstitialDismissed(InterstitialPresenter interstitialPresenter) {
+        invokeOnDismissed();
+    }
+}
